requirements.txt
Django==1.10
django-cart==1.0.4
django-datatable==0.3.1
django-hitcount==1.2.4
djangorestframework==3.7.3
mysqlclient==1.3.10
olefile==0.44
Pillow==4.3.0
